Only display backdrops from active mods
A bit hacky, but simple: stores the assembly name as well as a boolean
stating whether the backdrop should be rendered or not.
If Everest has not loaded the assembly in question, the setting is not
shown in the menu.

However, the setting is kept in case the mod is re-enabled in the
future.