Major UI revamp: move everything to submenus
... including the "Backdrops From Mods" submenu, which is now visible
in-game, and dynamically adds every backdrop seen!

This has been a LOT of work, most of which was ultimately only needed
for me to learn more about how Everest works.

I rewrote the settings system from scratch based on ExtendedVariantMode,
then while "porting" everything I realized I either need to find a way
to automatically populate the menu based on the properties in the
Settings file, or I needed code to store every settings in the Settings
file **and** code to create the menu for that setting in a separate
file, which felt too ugly.

I copied over the code Everest uses to populate the menu, modified it a
bit, got it to work, and then realized none of the changes I made to the
Everest code was actually required, and everything worked with
undocumented (but supported) Everest features!