Add centered wind indicator setting; remove offset
Offset was mostly left in to allow people to move it to the center
manually, but that's just an ugly hack.
There's little reason to adjust the offset if the indicator is properly
placed to begin with.

Also cleans up two subtexts that didn't do any good after the menu
options were renamed.