Fix Speedrun Tool compatibility w/ screen wipe patch
This mod was incompatible with the "Speedrun Tool" mod prior to this
change, as it uses the onComplete action on screen wipes to do several
important things, and this mod ignored screen wipes *and* their
onComplete actions.

After this commit, onComplete actions are executed just prior to the
next Update cycle even when screen wipes are disabled, which works with
Speedrun Tool.
When screen wipes are enabled (i.e. not modded) they execute the action
as usual.

Not tested with other mods that require these onComplete actions as I'm
not aware of any.